pyproject.toml
[tool.poetry]
name = "toxicity-prediction"
version = "0.1.0"
description = "Toxicity prediction for chemical compounds using MLP and GCN models"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "toxicity_prediction"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.0.0"
torch-geometric = "^2.4.0"
pytorch-lightning = "^2.1.0"
hydra-core = "^1.3.0"
omegaconf = "^2.3.0"
fire = "^0.5.0"
mlflow = "^2.9.0"
dvc = "^3.30.0"
pandas = "^1.5.0"
numpy = "^1.24.0"
rdkit = "^2024.0.0"
onnx = "^1.15.0"
onnxruntime = "^1.16.0"
onnxscript = "^0.5.0"
scikit-learn = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
black = "^23.12.0"
isort = "^5.13.0"
flake8 = "^6.1.0"

[tool.poetry.scripts]
toxicity = "toxicity_prediction.commands:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

